bib: add new `cntdnf` go module and move container dnf logic in there
This commit moves the logic of dealing with `osbuild-json-dnf` for
containers into a new container dnf `cntdnf` go module. Also move
the integration test for dnfjson with subscribed/normal content
there.
