+++ bib/cmd/bootc-image-builder/image.go
@@ -60,9 +60,6 @@ type ManifestConfig struct {
 	// Extracted information about the source container image
 	SourceInfo *source.Info
 
-	// Command to run the depsolver
-	DepsolverCmd []string
-
 	// RootFSType specifies the filesystem type for the root partition
 	RootFSType string
 }
