+++ bib/internal/container/container_test.go
@@ -8,22 +8,13 @@ import (
 	"os/exec"
 	"path"
 	"path/filepath"
-	"runtime"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
-
-	"github.com/osbuild/images/pkg/dnfjson"
-	"github.com/osbuild/images/pkg/rpmmd"
-
-	"github.com/osbuild/bootc-image-builder/bib/internal/source"
 )
 
-const (
-	testingImage    = "registry.access.redhat.com/ubi9-micro:latest"
-	dnfTestingImage = "registry.access.redhat.com/ubi9:latest"
-)
+const testingImage = "registry.access.redhat.com/ubi9-micro:latest"
 
 type containerInfo struct {
 	State string `json:"State"`
@@ -179,95 +170,4 @@ echo '%s'
 		_, err := cnt.DefaultRootfsType()
 		assert.ErrorContains(t, err, "unsupported root filesystem type: ext1, supported: ")
 	}
-}
-
-func subscribeMachine(t *testing.T) (restore func()) {
-	if _, err := exec.LookPath("subscription-manager"); err != nil {
-		t.Skip("no subscription-manager found")
-		return func() {}
-	}
-
-	matches, err := filepath.Glob("/etc/pki/entitlement/*.pem")
-	if err == nil && len(matches) > 0 {
-		return func() {}
-	}
-
-	rhsmOrg := os.Getenv("RHSM_ORG")
-	rhsmActivationKey := os.Getenv("RHSM_ACTIVATION_KEY")
-	if rhsmOrg == "" || rhsmActivationKey == "" {
-		t.Skip("no RHSM_{ORG,ACTIVATION_KEY} env vars found")
-		return func() {}
-	}
-
-	err = exec.Command("subscription-manager", "register",
-		"--org", rhsmOrg,
-		"--activationkey", rhsmActivationKey).Run()
-	require.NoError(t, err)
-
-	return func() {
-		exec.Command("subscription-manager", "unregister").Run()
-	}
-}
-
-func TestDNFInitGivesAccessToSubscribedContent(t *testing.T) {
-	if os.Geteuid() != 0 {
-		t.Skip("skipping test; not running as root")
-	}
-	if runtime.GOARCH != "amd64" {
-		t.Skip("skipping test; only runs on x86_64")
-	}
-
-	restore := subscribeMachine(t)
-	defer restore()
-
-	cnt, err := New(dnfTestingImage)
-	require.NoError(t, err)
-	err = cnt.InitDNF()
-	require.NoError(t, err)
-
-	content, err := cnt.ReadFile("/etc/yum.repos.d/redhat.repo")
-	require.NoError(t, err)
-	assert.Contains(t, string(content), "rhel-9-for-x86_64-baseos-rpms")
-}
-
-// XXX: should tihs be in a different file, it's more an integration test
-func TestDNFJsonWorkWithSubscribedContent(t *testing.T) {
-	if os.Geteuid() != 0 {
-		t.Skip("skipping test; not running as root")
-	}
-	if runtime.GOARCH != "amd64" {
-		t.Skip("skipping test; only runs on x86_64")
-	}
-	if _, err := os.Stat("/usr/libexec/osbuild-depsolve-dnf"); err != nil {
-		t.Skip("cannot find /usr/libexec/osbuild-depsolve-dnf")
-	}
-	cacheRoot := t.TempDir()
-
-	restore := subscribeMachine(t)
-	defer restore()
-
-	cnt, err := New(dnfTestingImage)
-	require.NoError(t, err)
-	err = cnt.InitDNF()
-	require.NoError(t, err)
-	depsolverCmd, err := cnt.InitDepSolveDNF()
-	require.NoError(t, err)
-
-	sourceInfo, err := source.LoadInfo(cnt.Root())
-	require.NoError(t, err)
-	solver := dnfjson.NewSolver(
-		sourceInfo.OSRelease.PlatformID,
-		sourceInfo.OSRelease.VersionID,
-		"x86_64",
-		fmt.Sprintf("%s-%s", sourceInfo.OSRelease.ID, sourceInfo.OSRelease.VersionID),
-		cacheRoot)
-	solver.SetDNFJSONPath(depsolverCmd[0], depsolverCmd[1:]...)
-	solver.SetRootDir("/")
-	res, err := solver.Depsolve([]rpmmd.PackageSet{
-		{
-			Include: []string{"coreutils"},
-		},
-	}, 0)
-	require.NoError(t, err)
-	assert.True(t, len(res.Packages) > 0)
 }
